chore: Fix RuboCop issue
